課題作成機能実装（問題選択・抽出機能作成中）

+++ src/main/java/jp/ac/ems/repository/TaskRepository.java
@@ -8,6 +8,6 @@ import jp.ac.ems.bean.TaskBean;
  * 課題用リポジトリ(taks repository).
  * @author tejc999999
  */
-public interface TaskRepository extends JpaRepository<TaskBean, Integer> {
+public interface TaskRepository extends JpaRepository<TaskBean, Long> {
 
 }
